Fix Rails 6 deprecation warning
This addresses the following:

```
DEPRECATION WARNING: Using a dynamic :action segment in a route is deprecated and will be removed in Rails 6.0.
```

Since the routes are only used in one describe block, I find for tests that keeping that as close as possible to where its used is best rather than having it globally. It makes it easier to refactor/change later if necessary.