Remove lint rake task
This is no longer needed as we migrated the single lint spec to a normal spec. The lint for the carriage return isn't necessary and rather not have more code to maintain and use something off the shelf. Primary code here is Ruby and Rubocop already handles this lint check for us.