Add Codecov setup (#469)
* Add simplecov gem

* Rename group to rubocop

* Create codecov.yml

* Add simplecov setup to spec_helper

* Migrate gemspec_spec.lint.rb to regular spec

No need to treat this as half lint half spec. Just make it a spec and keep it simple.

* Remove lint rake task

This is no longer needed as we migrated the single lint spec to a normal spec. The lint for the carriage return isn't necessary and rather not have more code to maintain and use something off the shelf. Primary code here is Ruby and Rubocop already handles this lint check for us.

* Enable Rubocop Layout/TrailingEmptyLines

* Enable Rubocop Layout/ExtraSpacing

* Enable Rubocop Layout/TrailingWhitespace

* Remove duplicate ruby setup in CI

* Add code coverage upload to CI