bin/bundle update --bundler
This bumps locked bundler version to the same we are using on ActiveAdmin and InheritedResources. For Arbre PR's I noticed the Rubocop job installing many gems when it shouldn't have been. The bundler version was behind what we needed for the workflow change in https://github.com/activeadmin/activeadmin-github-workflows/pull/3