Include blank attributes in HTML output (#517)
This is mostly a full revert of commit 15267891a4c741154c7ee601d86851b16516588a which was a bad change. Empty attributes are valid HTML and should be supported in Arbre (e.g. boolean attributes). The bad commit was for just addressing the class attribute but that applied to any attribute in HTML when it shouldn't have, even for class.