Include empty attributes in HTML output
This is mostly a full revert of commit 1526789 which was a bad change. Empty attributes are valid HTML and should be supported in Arbre (e.g. boolean attributes). The bad commit was for just addressing the class attribute but that applied to any attribute in HTML when it shouldn't have, even for class. Note that if an attribute value is nil then it is removed. This follows similar output logic to Rails `tag.attributes` helper.