Ensure Arbre context is reset
When you have an Arbre component with a block form, it pushes the current
node onto a stack so that it can use [method_missing][1] to provide a dynamic
evaluation context (eg. so tab is available only directly inside of tabs).

Unfortunately, when an error is raised inside inside one of these, if we
have a rescue to catch that error so the rest of the page can be rendered -
the context is still stuck inside the last block that was added. Other than
likely rendering some of the wrong markup (depending on usage), it means
the context is not put back into the block where the rescue is located.
As a result, even though that block is effectively thrown away when a
rescue is located above the erroring call, certain features are not available.

[1]: https://github.com/activeadmin/arbre/blob/v2.2.0/lib/arbre/element.rb#L176-L186